command: s/givenCommand/commandFromContext
